+++ exercises/filtered_lsls/exercise.js
@@ -1,15 +1,16 @@
 var fs            = require('fs')
   , path          = require('path')
   , os            = require('os')
-  , rimraf        = require('rimraf')
   , exercise      = require('workshopper-exercise')()
   , filecheck     = require('workshopper-exercise/filecheck')
   , execute       = require('workshopper-exercise/execute')
   , comparestdout = require('workshopper-exercise/comparestdout')
   , wrappedexec   = require('workshopper-wrappedexec')
-  , boganipsum    = require('boganipsum')
+  , after         = require('after')
+  , rimraf        = require('rimraf')
+  , files         = require('./file-list')
 
-  , testFile      = path.join(os.tmpDir(), '_learnyounode_' + process.pid + '.txt')
+  , testDir       = path.join(os.tmpDir(), '_learnyounode_' + process.pid)
 
 
 // checks that the submission file actually exists
@@ -27,25 +28,37 @@ exercise = wrappedexec(exercise)
 
 // a module we want run just prior to the submission in the
 // child process
-exercise.wrapModule(require.resolve('../my_first_io/wrap'))
+exercise.wrapModule(require.resolve('../my_first_ioio/wrap'))
 
 
 // set up the data file to be passed to the submission
 exercise.addSetup(function (mode, callback) {
   // mode == 'run' || 'verify'
 
-  var lines = Math.ceil(Math.random() * 50)
-    , txt   = boganipsum({ paragraphs: lines })
-
-  // supply the file as an arg to the 'execute' processor for both
+  // supply the dir and extensions as args to the 'execute' processor for both
   // solution and submission spawn()
   // using unshift here because wrappedexec needs to use additional
   // args to do its magic
-  this.submissionArgs.unshift(testFile)
-  this.solutionArgs.unshift(testFile)
+  this.submissionArgs.unshift('md')
+  this.submissionArgs.unshift(testDir)
+  this.solutionArgs.unshift('md')
+  this.solutionArgs.unshift(testDir)
 
-  // file with random text
-  fs.writeFile(testFile, txt, 'utf8', callback)
+  fs.mkdir(testDir, function (err) {
+    if (err)
+      return callback(err)
+
+    var done = after(files.length, callback)
+
+    files.forEach(function (f) {
+      fs.writeFile(
+          path.join(testDir, f)
+        , 'nothing to see here'
+        , 'utf8'
+        , done
+      )
+    })
+  })
 })
 
 
@@ -64,6 +77,9 @@ exercise.addVerifyProcessor(function (callback) {
     }
   }.bind(this))
 
+  if (!usedSync && !usedAsync) // https://github.com/nodeschool/discussions/issues/356
+    this.emit('fail', 'Used asynchronous method from the `fs` module')
+
   callback(null, usedAsync && !usedSync)
 })
 
@@ -72,7 +88,7 @@ exercise.addVerifyProcessor(function (callback) {
 exercise.addCleanup(function (mode, passed, callback) {
   // mode == 'run' || 'verify'
 
-  rimraf(testFile, callback)
+  rimraf(testDir, callback)
 })
