+++ exercises/http_uppercasererhttp_/exercise.js
