+++ exercises/menu.json
@@ -1,15 +1,15 @@
 [
-    "こんにちは世界"
-  , "ベイビーステップ"
-  , "私の初めてのI/O！"
-  , "私の初めてのASYNC I/O！"
-  , "LSのフィルター"
-  , "モジュラーに作ってください"
-  , "HTTP クライエント"
-  , "HTTP 集める"
-  , "ASYNCジャグリング"
-  , "タイムサーバ"
-  , "HTTP ファイルサーバ"
-  , "HTTP アッパーケース"
-  , "HTTP JSON API サーバ"
+    "HELLO WORLD（こんにちは世界）"
+  , "BABY STEPS（ベイビーステップ）"
+  , "MY FIRST I/O!（私の初めてのI/O！）"
+  , "MY FIRST ASYNC I/O!（私の初めてのASYNC I/O！）"
+  , "FILTERED LS（LSのフィルター）"
+  , "MAKE IT MODULAR（モジュラーにしましょう）"
+  , "HTTP CLIENT（HTTP クライエント）"
+  , "HTTP COLLECT（HTTP 集める）"
+  , "JUGGLING ASYNC（ASYNCジャグリング）"
+  , "TIME SERVER（タイムサーバ）"
+  , "HTTP FILE SERVER（HTTP ファイルサーバ）"
+  , "HTTP UPPERCASERER（HTTP アッパーケース）"
+  , "HTTP JSON API SERVER（HTTP JSON API サーバ）"
 ]
